Refactor ApiRequest
