+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/apirequest/HavingGenerator.java
@@ -1,6 +1,5 @@
 // Copyright 2017 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-
 package com.yahoo.bard.webservice.web.apirequest;
 
 import com.yahoo.bard.webservice.data.metric.LogicalMetric;
