+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/DimensionApiRequestMapperSpec.groovy
@@ -3,6 +3,7 @@
 package com.yahoo.bard.webservice.web
 import com.yahoo.bard.webservice.application.JerseyTestBinder
 import com.yahoo.bard.webservice.data.config.ResourceDictionaries
+import com.yahoo.bard.webservice.web.apirequest.DimensionsApiRequestImpl
 import com.yahoo.bard.webservice.web.endpoints.DimensionsServlet
 
 import spock.lang.Specification
@@ -24,12 +25,12 @@ class DimensionApiRequestMapperSpec extends Specification {
         setup:
         ResourceDictionaries resourceDictionaries = jtb.configurationLoader.getDictionaries()
         DimensionApiRequestMapper dimensionApiRequestMapper = new DimensionApiRequestMapper(resourceDictionaries)
-        DimensionsApiRequest dimensionsApiRequest
+        DimensionsApiRequest apiRequest
         DimensionsApiRequest mappedDimensionApiRequest
         Exception e
 
         when:
-        dimensionsApiRequest = new DimensionsApiRequest(
+        apiRequest = new DimensionsApiRequestImpl(
                 "color",
                 "shape|desc-in[shape]",
                 null,
@@ -38,14 +39,14 @@ class DimensionApiRequestMapperSpec extends Specification {
                 resourceDictionaries.getDimensionDictionary(),
                 null
         )
-        dimensionApiRequestMapper.apply(dimensionsApiRequest, Mock(ContainerRequestContext))
+        dimensionApiRequestMapper.apply(apiRequest, Mock(ContainerRequestContext))
 
         then:
         e = thrown(BadApiRequestException)
         e.getMessage().matches("Filter.*not match dimension.*")
 
         when:
-        dimensionsApiRequest = new DimensionsApiRequest(
+        apiRequest = new DimensionsApiRequestImpl(
                 "color",
                 "shape|desc-in[shape],color|desc-in[red]",
                 null,
@@ -54,14 +55,14 @@ class DimensionApiRequestMapperSpec extends Specification {
                 resourceDictionaries.getDimensionDictionary(),
                 null
         )
-        dimensionApiRequestMapper.apply(dimensionsApiRequest, Mock(ContainerRequestContext))
+        dimensionApiRequestMapper.apply(apiRequest, Mock(ContainerRequestContext))
 
         then:
         e = thrown(BadApiRequestException)
         e.getMessage().matches("Filter.*not match dimension.*")
 
         when:
-        dimensionsApiRequest = new DimensionsApiRequest(
+        apiRequest = new DimensionsApiRequestImpl(
                 "color",
                 "color|desc-in[red]",
                 null,
@@ -70,9 +71,9 @@ class DimensionApiRequestMapperSpec extends Specification {
                 resourceDictionaries.getDimensionDictionary(),
                 null
         )
-        mappedDimensionApiRequest = dimensionApiRequestMapper.apply(dimensionsApiRequest, Mock(ContainerRequestContext))
+        mappedDimensionApiRequest = dimensionApiRequestMapper.apply(apiRequest, Mock(ContainerRequestContext))
 
         then:
-        mappedDimensionApiRequest == dimensionsApiRequest
+        mappedDimensionApiRequest == apiRequest
     }
 }
